// src/app/enums/AddMemberDialogTypeEnum.ts
export enum AddMemberDialogTypeEnum {
  addMemberToTeam = 'addMemberToTeam',
  addMemberToDepartment = 'addMemberToDepartment',
  addMemberToProject = 'addMemberToProject',
  addMemberToTask = 'addMemberToTask',
  inviteMemberToEvent = 'inviteMemberToEvent',
  addMemberToReminder = 'addMemberToReminder'
}